+++ src/main/java/com/platzi/market/domain/repository/ProductRepository.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface ProductRepository {
     List<Product> getAll();
-    Optional<List<Product>> getByCategory(long categoryId);
-    Optional<List<Product>> getScarseProducts(long quantity);
-    Optional<Product> getProduct(long productId);
+    Optional<List<Product>> getByCategory(int categoryId);
+    Optional<List<Product>> getScarseProducts(int quantity);
+    Optional<Product> getProduct(int productId);
     Product save(Product product);
-    void delete(long productId);
+    void delete(int productId);
 }
