+++ src/main/java/com/platzi/market/persistence/crud/ProductoCrudRepository.java
@@ -7,13 +7,12 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductoCrudRepository extends CrudRepository <Producto, Long> {
+public interface ProductoCrudRepository extends CrudRepository <Producto, Integer> {
 
     /* MANERA NATIVA */
     //@Query(value = "SELECT * FROM productos WHERE id_categoria = ?", nativeQuery = true)
 
     /* MANERA CON QUERY METHODS */
-    List<Producto> findByIdCategoriaOrderByNombreAsc(long idCategoria);
-
-    Optional<List<Producto>> findByCantidadStockLessThanAndEstado(long cantidadStock, boolean estado);
+    List<Producto> findByIdCategoriaOrderByNombreAsc(int idCategoria);
+    Optional<List<Producto>> findByCantidadStockLessThanAndEstado(int cantidadStock, boolean estado);
 }
